rename `region_inference` module to `region_constraints`
